+++ NEWS.md
@@ -1,8 +1,9 @@
 # mpg2wav (development version)
 
-* Added support for GitHub
+* Added dependency from `{av}` for audio-video transformations.
+* Added support for GitHub.
 
 # mpg2wav 0.0.0.9000
 
-* Added basic dev infrastructure
+* Added basic dev infrastructure.
 * Added a `NEWS.md` file to track changes to the package.